chg 文件夹改名，给datasource不进行默认加载
