+++ src/test/java/com/simonalong/neo/db/BaseNeoTableTest.java
@@ -1,4 +1,4 @@
-package com.simonalong.neo.table;
+package com.simonalong.neo.db;
 
 import com.simonalong.neo.BaseTest;
 import com.simonalong.neo.Neo;
