+++ src/test/java/com/simonalong/neo/db/NeoTableTest.java
@@ -1,4 +1,4 @@
-package com.simonalong.neo.table;
+package com.simonalong.neo.db;
 
 import com.simonalong.neo.Columns;
 import com.simonalong.neo.NeoMap;
