+++ src/test/java/com/simonalong/neo/neo/NeoDemo.java
@@ -3,8 +3,8 @@ package com.simonalong.neo.neo;
 import com.simonalong.neo.Neo;
 import com.simonalong.neo.NeoMap;
 import com.simonalong.neo.entity.DemoEntity3;
-import com.simonalong.neo.table.NeoPage;
-import com.simonalong.neo.table.NeoTable;
+import com.simonalong.neo.db.NeoPage;
+import com.simonalong.neo.db.NeoTable;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.Test;
