+++ src/test/java/com/simonalong/neo/neo/NeoPageTest.java
@@ -3,7 +3,7 @@ package com.simonalong.neo.neo;
 import com.simonalong.neo.Columns;
 import com.simonalong.neo.NeoBaseTest;
 import com.simonalong.neo.NeoMap;
-import com.simonalong.neo.table.NeoPage;
+import com.simonalong.neo.db.NeoPage;
 import com.simonalong.neo.entity.DemoEntity;
 import java.sql.SQLException;
 import lombok.SneakyThrows;
